Refactor: Type 컴포넌트명 GatheringType으로 변경
